[MISC] move exposition_only header
