+++ include/seqan3/alignment/matrix/detail/affine_cell_proxy.hpp
@@ -18,7 +18,7 @@
 
 #include <seqan3/alignment/matrix/detail/trace_directions.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/tuple/concept.hpp>
