+++ include/seqan3/alignment/matrix/detail/alignment_score_matrix_proxy.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 
 namespace seqan3::detail
