+++ include/seqan3/alignment/scoring/scoring_scheme_base.hpp
@@ -18,7 +18,7 @@
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 #if SEQAN3_WITH_CEREAL
 #include <cereal/types/array.hpp>
