+++ include/seqan3/io/detail/misc_output.hpp
@@ -26,7 +26,7 @@
     #include <seqan3/contrib/stream/gz_ostream.hpp>
 #endif
 #include <seqan3/io/exception.hpp>
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 namespace seqan3::detail
 {
