+++ include/seqan3/io/sam_file/sam_tag_dictionary.hpp
@@ -18,8 +18,8 @@
 
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 #include <seqan3/utility/container/small_string.hpp>
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
 
 namespace seqan3::detail
 {
