+++ include/seqan3/search/configuration/max_error_common.hpp
@@ -16,7 +16,7 @@
 #include <seqan3/std/concepts>
 
 #include <seqan3/core/detail/strong_type.hpp>
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 namespace seqan3::search_cfg
 {
