+++ include/seqan3/utility/simd/views/iota_simd.hpp
@@ -14,7 +14,7 @@
 
 #include <seqan3/std/ranges>
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
