+++ test/snippet/alphabet/nucleotide/@target_alphabet@_implicit_conversion_from_@source_alphabet@_inherit.cpp.in
@@ -21,7 +21,7 @@ int main()
 }
 //![main]
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 static_assert(seqan3::implicitly_convertible_to<seqan3::${source_alphabet}, seqan3::${target_alphabet}>);
 static_assert(!seqan3::implicitly_convertible_to<seqan3::${source_alphabet}, my_${target_alphabet}>);
