+++ test/snippet/alphabet/nucleotide/dna15_implicit_conversion_from_rna15_inherit.cpp
@@ -21,7 +21,7 @@ int main()
 }
 //![main]
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 static_assert(seqan3::implicitly_convertible_to<seqan3::rna15, seqan3::dna15>);
 static_assert(!seqan3::implicitly_convertible_to<seqan3::rna15, my_dna15>);
