+++ test/snippet/alphabet/nucleotide/dna4_implicit_conversion_from_rna4_inherit.cpp
@@ -21,7 +21,7 @@ int main()
 }
 //![main]
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 static_assert(seqan3::implicitly_convertible_to<seqan3::rna4, seqan3::dna4>);
 static_assert(!seqan3::implicitly_convertible_to<seqan3::rna4, my_dna4>);
