+++ test/snippet/alphabet/nucleotide/rna15_implicit_conversion_from_dna15_inherit.cpp
@@ -21,7 +21,7 @@ int main()
 }
 //![main]
 
-#include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/concept/exposition_only/core_language.hpp>
 
 static_assert(seqan3::implicitly_convertible_to<seqan3::dna15, seqan3::rna15>);
 static_assert(!seqan3::implicitly_convertible_to<seqan3::dna15, my_rna15>);
